Implement the 'APIRouter' to organize the 'main.py' file
